# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pokemon Anil API",
    "version": "2.0.1",
    "description": "API propria do Pokemon Anil/PT-BR. Dados locais Azul 4.0.6 (interna 4.0.3). Sem consultas externas nos catalogos de jogo."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/pokemon": {
      "get": {
        "summary": "GET /pokemon",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ability",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "generation",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "min_bst",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/types": {
      "get": {
        "summary": "GET /types",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        }
      }
    },
...
    "/anil": {
      "get": {
        "summary": "GET /anil",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        }
      }
    },
    "/anil/modes": {
      "get": {
        "summary": "GET /anil/modes",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        }
      }
    },
    "/anil/modes/{id}": {
      "get": {
        "summary": "GET /anil/modes/{id}",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/anil/mechanics": {
      "get": {
        "summary": "GET /anil/mechanics",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        }
      }
    },
    "/anil/mechanics/{id}": {
      "get": {
        "summary": "GET /anil/mechanics/{id}",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/anil/evolution-changes": {
      "get": {
        "summary": "GET /anil/evolution-changes",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        }
      }
    },
    "/anil/evolution-changes/{pokemon}": {
      "get": {
        "summary": "GET /anil/evolution-changes/{pokemon}",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        },
        "parameters": [
          {
            "name": "pokemon",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/anil/key-items": {
      "get": {
        "summary": "GET /anil/key-items",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        }
      }
    },
...
    "/anil/locations": {
      "get": {
        "summary": "GET /anil/locations",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        }
      }
    },
...
    "/anil/rates": {
      "get": {
        "summary": "GET /anil/rates",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        }
      }
    },
...
    "/abilities": {
      "get": {
        "summary": "GET /abilities",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PokeAPI data"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/items": {
      "get": {
        "summary": "GET /items",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PokeAPI data"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/anil/trainers": {
      "get": {
        "summary": "Treinadores e equipes da compila??o",
        "responses": {
          "200": {
            "description": "Dados locais do jogo"
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ]
      }
    },
...
    "/moves": {
      "get": {
        "summary": "Cat?logo de golpes presentes no jogo",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Golpes da compila??o local"
          },
          "400": {
            "description": "Pagina??o inv?lida"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "writeKey": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}